requirements.txt
pip
setuptools
numpy
PyYAML
matplotlib
tensorflow-gpu==2.10